api/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-multipart
minio
email-validator
